refactor: remove unused apiVersion from getDefinition and delete globalTestHelper
getDefinition never used the apiVersion parameter from its config type.
Remove it, replace getGlobalMetadata() with direct getDefinition({}) calls
in tests, move getWork() to dedicated testWork.ts, and fix 3 tests that
had hidden dependencies on setupFiles pre-loading @salesforce/core.
