+++ __tests__/unit/lib/service/customObjectHandler.test.ts
@@ -3,6 +3,7 @@ import { describe, expect, it, jest } from '@jest/globals'
 
 import { MASTER_DETAIL_TAG } from '../../../../src/constant/metadataConstants'
 import { MetadataRepository } from '../../../../src/metadata/MetadataRepository'
+import { getDefinition } from '../../../../src/metadata/metadataManager'
 import CustomObjectHandler from '../../../../src/service/customObjectHandler'
 import type { Work } from '../../../../src/types/work'
 import {
@@ -11,7 +12,7 @@ import {
   readDirs,
   readPathFromGit,
 } from '../../../../src/utils/fsHelper'
-import { getGlobalMetadata, getWork } from '../../../__utils__/globalTestHelper'
+import { getWork } from '../../../__utils__/testWork'
 
 jest.mock('../../../../src/utils/fsHelper')
 
@@ -61,7 +62,7 @@ beforeEach(() => {
 describe('CustomObjectHandler', () => {
   let globalMetadata: MetadataRepository
   beforeAll(async () => {
-    globalMetadata = await getGlobalMetadata()
+    globalMetadata = await getDefinition({})
   })
 
   describe('when called with generateDelta false', () => {
