+++ __tests__/unit/lib/service/diffLineInterpreter.test.ts
@@ -2,9 +2,10 @@
 import { describe, expect, it, jest } from '@jest/globals'
 
 import { MetadataRepository } from '../../../../src/metadata/MetadataRepository'
+import { getDefinition } from '../../../../src/metadata/metadataManager'
 import DiffLineInterpreter from '../../../../src/service/diffLineInterpreter'
 import type { Work } from '../../../../src/types/work'
-import { getGlobalMetadata, getWork } from '../../../__utils__/globalTestHelper'
+import { getWork } from '../../../__utils__/testWork'
 
 const mockHandle = jest.fn()
 jest.mock('../../../../src/service/typeHandlerFactory', () => {
@@ -29,7 +30,7 @@ describe('DiffLineInterpreter', () => {
   let sut: DiffLineInterpreter
   let globalMetadata: MetadataRepository
   beforeAll(async () => {
-    globalMetadata = await getGlobalMetadata()
+    globalMetadata = await getDefinition({})
   })
 
   beforeEach(() => {
