+++ __tests__/unit/lib/service/diffLineInterpreterCompatibility.test.ts
@@ -3,11 +3,13 @@
 import { describe, expect, it, jest } from '@jest/globals'
 
 import { MetadataRepository } from '../../../../src/metadata/MetadataRepository'
+import { getDefinition } from '../../../../src/metadata/metadataManager'
 import DiffLineInterpreter from '../../../../src/service/diffLineInterpreter'
 import type { Work } from '../../../../src/types/work'
-import { getGlobalMetadata, getWork } from '../../../__utils__/globalTestHelper'
+import { getWork } from '../../../__utils__/testWork'
 
 jest.mock('node:os', () => ({
+  ...(jest.requireActual('node:os') as object),
   availableParallelism: null,
 }))
 
@@ -34,7 +36,7 @@ describe('DiffLineInterpreter', () => {
   let sut: DiffLineInterpreter
   let globalMetadata: MetadataRepository
   beforeAll(async () => {
-    globalMetadata = await getGlobalMetadata()
+    globalMetadata = await getDefinition({})
   })
 
   describe('compatibility test', () => {
