+++ __tests__/unit/lib/utils/MessageService.test.ts
@@ -13,6 +13,14 @@ jest.mock('@salesforce/core', () => {
         getMessages: mockedMessages,
       })),
     },
+    Logger: {
+      childFromRoot: jest.fn(() => ({
+        setLevel: jest.fn(),
+        shouldLog: jest.fn(),
+        trace: jest.fn(),
+      })),
+    },
+    LoggerLevel: {},
   }
 })
